src/api/requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydobot==1.3.2
PyMySQL==1.0.2
pyserial==3.4
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
Werkzeug==2.2.3
